+++ lib/permissions/index.ts
@@ -1,1 +1,3 @@
 export * from "./rolePermissions";
+
+export * from "./roleRelationshipMap";